@@ -11,18 +11,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.12",
-    "@angular/cdk": "17.3.10",
-    "@angular/common": "17.3.12",
-    "@angular/compiler": "17.3.12",
-    "@angular/core": "17.3.12",
-    "@angular/forms": "17.3.12",
-    "@angular/material": "17.3.10",
-    "@angular/material-luxon-adapter": "17.3.10",
-    "@angular/platform-browser": "17.3.12",
-    "@angular/platform-browser-dynamic": "17.3.12",
-    "@angular/router": "17.3.12",
-    "@angular/youtube-player": "17.3.10",
+    "@angular/animations": "18.1.2",
+    "@angular/cdk": "18.1.2",
+    "@angular/common": "18.1.2",
+    "@angular/compiler": "18.1.2",
+    "@angular/core": "18.1.2",
+    "@angular/forms": "18.1.2",
+    "@angular/material": "18.1.2",
+    "@angular/material-luxon-adapter": "18.1.2",
+    "@angular/platform-browser": "18.1.2",
+    "@angular/platform-browser-dynamic": "18.1.2",
+    "@angular/router": "18.1.2",
+    "@angular/youtube-player": "18.1.2",
     "file-saver": "2.0.5",
     "get-video-id": "3.7.0",
     "luxon": "3.4.4",
@@ -31,22 +31,21 @@
     "zone.js": "0.14.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.8",
-    "@angular-eslint/builder": "17.5.2",
-    "@angular-eslint/eslint-plugin": "17.5.2",
-    "@angular-eslint/eslint-plugin-template": "17.5.2",
-    "@angular-eslint/schematics": "17.5.2",
-    "@angular-eslint/template-parser": "17.5.2",
-    "@angular/cli": "17.3.8",
-    "@angular/compiler-cli": "17.3.12",
+    "@angular-devkit/build-angular": "18.1.2",
+    "@angular-eslint/builder": "18.1.0",
+    "@angular-eslint/eslint-plugin": "18.1.0",
+    "@angular-eslint/eslint-plugin-template": "18.1.0",
+    "@angular-eslint/schematics": "18.1.0",
+    "@angular-eslint/template-parser": "18.1.0",
+    "@angular/cli": "18.1.2",
+    "@angular/compiler-cli": "18.1.2",
     "@types/file-saver": "2.0.7",
     "@types/luxon": "3.4.2",
     "@types/node": "20.14.13",
-    "@typescript-eslint/eslint-plugin": "7.18.0",
-    "@typescript-eslint/parser": "7.18.0",
-    "eslint": "8.57.0",
-    "eslint-plugin-deprecation": "2.0.0",
+    "@typescript-eslint/eslint-plugin": "8.0.0-alpha.58",
+    "@typescript-eslint/parser": "8.0.0-alpha.58",
+    "eslint": "9.8.0",
     "schematics-scss-migrate": "2.3.17",
-    "typescript": "5.4.5"
+    "typescript": "5.5.4"
   }
 }
