@@ -1,9 +1,9 @@
 import {Component} from '@angular/core';
 
 @Component({
-  selector: 'app-game-discussion-finish-comments-dialog',
-  templateUrl: './game-discussion-finish-comments-dialog.component.html',
-  styleUrls: ['./game-discussion-finish-comments-dialog.component.scss']
+    selector: 'app-game-discussion-finish-comments-dialog',
+    templateUrl: './game-discussion-finish-comments-dialog.component.html',
+    styleUrls: ['./game-discussion-finish-comments-dialog.component.scss']
 })
 export class GameDiscussionFinishCommentsDialogComponent {
 
