@@ -35,7 +35,9 @@ export class AuthenticationInterceptor implements HttpInterceptor {
         if (err.status === 401 || err.status === 403) {
             this.authenticationService.logout();
             if (!this.router.url.endsWith(LOGIN_PATH)) {
-                this.router.navigate([LOGIN_PATH]);
+                this.router.navigate([LOGIN_PATH]).catch(reason => {
+                    console.error(reason);
+                });
                 return of(err.message);
             }
         }
